update OrdersService deployment for EF migrations
Add a new strategy object to deployment.yaml that specifies that all replicas are terminated (maxUnavailable: 100%)
before new replicas are deployed one at a time (maxSurge: 1).
This ensures that EF Core migrations can properly run in a single pod.
When the readiness probe suceeds, the next replica is started which
sees the target database is up-to-date.