Admin CLI (#117)
* add new client to authentication service for admin CLI

This commit adds a new client to identity server with the client credentials flow. It can be used by the admin CLI tool which is yet to be created.

* add PriceDropNotifier.csproj to .NET sln

* admin-cli client now also has the sample scope

* add admin CLI project

This commit contains the basic structure of the admin CLI project with logging, transparent authentication with the client credentials flow, secure persistence of app settings, and an easily extensible CLI command structure

* add PriceWatcher csproj to sln

* add DropPriceCommand to admin CLI

* help is now shown when admin CLI is called without parameters

* add change log level command to admin CLI

* add TargetCloudCommand and TargetLocalCommand to admin CLI

* add missing descriptions to DropPriceCommand and ListProductsCommand in admin CLI

* fix error in AppSettings.TryGetAccessToken

* rename jwtSecurityToken to token

* changed CLI command structure

The commands now use "products" and "environment" on the top layer with various subcommands to show or manipulate.

* change option IDs to kebab-case

* use double.TryParse in DropPriceCommand instead of custom number parsing

* introduce AccessTokenValidToThreshold

Previously, the amount of seconds that was subtracted from the current access token's valid-to timestamp was hardcoded to 30 seconds. The AccessTokenValidToThreshold can now be set with environment variables and defaults to ten seconds.

* replace check whether products is null or empty in ListProductsCommand

* fixed invalid template in DropPriceCommand

* replace Serilog with plain Microsoft.Extensions.Logging

* admin-cli access token lifetime is now 4 hours

* remove dependency to Microsoft.Extensions.Configuration