feat: add ProductsAlignmentMigration
This migration inserts the same products that the in-memory repositories use to the Azure SQL database. Another great thing would be to actually retrieve the products in the price watcher service when checking that a product ID is valid.

I also restructured the migrations in this commit: Script as well as PostMigration are now executed within the same transaction.

Signed-off-by: Kenny Pflug <kenny.pflug@live.de>